+++ src/aim/models/applications.py
@@ -196,9 +196,10 @@ class CodePipeBuildDeploy(Resource):
 
 @implementer(schemas.IS3BucketPolicy)
 class S3BucketPolicy():
-    aws = FieldProperty(schemas.IS3BucketPolicy['aws'])
-    effect = FieldProperty(schemas.IS3BucketPolicy['effect'])
     action = FieldProperty(schemas.IS3BucketPolicy['action'])
+    aws = FieldProperty(schemas.IS3BucketPolicy['aws'])
+    condition = FieldProperty(schemas.IS3BucketPolicy['condition'])
+    effect = FieldProperty(schemas.IS3BucketPolicy['effect'])
     principal = FieldProperty(schemas.IS3BucketPolicy['principal'])
     resource_suffix = FieldProperty(schemas.IS3BucketPolicy['resource_suffix'])